ignore class load errors from the http server being shutdown
