+++ dd-java-agent/instrumentation/jsp-2.3/src/test/groovy/JSPInstrumentationForwardTests.groovy
@@ -1,80 +1,12 @@
-import com.google.common.io.Files
-import datadog.trace.agent.test.AgentTestRunner
-import datadog.trace.agent.test.utils.OkHttpUtils
-import datadog.trace.agent.test.utils.PortUtils
 import datadog.trace.api.DDSpanTypes
 import datadog.trace.bootstrap.instrumentation.api.Tags
-import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.Response
-import org.apache.catalina.Context
-import org.apache.catalina.startup.Tomcat
 import org.apache.jasper.JasperException
 import org.eclipse.jetty.http.HttpStatus
-import spock.lang.Shared
 import spock.lang.Unroll
 
-class JSPInstrumentationForwardTests extends AgentTestRunner {
-
-  @Shared
-  int port
-  @Shared
-  Tomcat tomcatServer
-  @Shared
-  Context appContext
-  @Shared
-  String jspWebappContext = "jsptest-context"
-
-  @Shared
-  File baseDir
-  @Shared
-  String baseUrl
-
-  OkHttpClient client = OkHttpUtils.client()
-
-  @Override
-  void configurePreAgent() {
-    super.configurePreAgent()
-    // skip jar scanning using environment variables:
-    // http://tomcat.apache.org/tomcat-7.0-doc/config/systemprops.html#JAR_Scanning
-    // having this set allows us to test with old versions of the tomcat api since
-    // JarScanFilter did not exist in the tomcat 7 api
-    System.setProperty("org.apache.catalina.startup.ContextConfig.jarsToSkip", "*")
-    System.setProperty("org.apache.catalina.startup.TldConfig.jarsToSkip", "*")
-  }
-
-  def setupSpec() {
-    baseDir = Files.createTempDir()
-    baseDir.deleteOnExit()
-
-    port = PortUtils.randomOpenPort()
-
-    tomcatServer = new Tomcat()
-    tomcatServer.setBaseDir(baseDir.getAbsolutePath())
-    tomcatServer.setPort(port)
-    tomcatServer.getConnector()
-    // comment to debug
-    tomcatServer.setSilent(true)
-    // this is needed in tomcat 9, this triggers the creation of a connector, will not
-    // affect tomcat 7 and 8
-    // https://stackoverflow.com/questions/48998387/code-works-with-embedded-apache-tomcat-8-but-not-with-9-whats-changed
-    tomcatServer.getConnector()
-
-    baseUrl = "http://localhost:$port/$jspWebappContext"
-
-    appContext = tomcatServer.addWebapp("/$jspWebappContext",
-      JSPInstrumentationForwardTests.getResource("/webapps/jsptest").getPath())
-
-    tomcatServer.start()
-    System.out.println(
-      "Tomcat server: http://" + tomcatServer.getHost().getName() + ":" + port + "/")
-  }
-
-  def cleanupSpec() {
-    tomcatServer.stop()
-    tomcatServer.destroy()
-  }
-
+class JSPInstrumentationForwardTests extends JSPTestBase {
   @Unroll
   def "non-erroneous GET forward to #forwardTo"() {
     setup:
